Fix stale endpoint detection in kube-proxy
There is an upstream bug in kube-proxy which can result in stale
conntrack entries, which in turn can result in failure of
communication between pods.

The upstream fix for kube-proxy was implemented in Kubernetes
commit 909925b4.  This fix has been backported to v1.21.8 in
kube-proxy-fix-stale-detection-logic.patch and a new kube-proxy
container image was generated containing the fix.

In order to actually make use of the new image, we need to modify
kubeadm to use the custom image tag rather than the default one.
This is implemented in
modify-kubeadm-to-use-custom-kube-proxy-image.patch.

NOTE1: This implementation will only affect platform upgrades to
k8s 1.21.8, it will not affect systems already running 22.12.

NOTE2: correction in 21.12 was done as a centos patch, this change
does it as a debian one

NOTE3: For upgrades from 21.12 to 22.12 the operator must ensure
that kube-proxy:v1.21.8-wrs.1 is available in the registry prior to
the upgrade process (use system registry-image-tags
k8s.gcr.io/kube-proxy to check that).

Test Plan:
PASS: Fresh install on AIO-SX
PASS: Fresh install on multinode lab
PASS: AIO-SX Upgrade from 21.12 to 22.12
PASS: AIO-DX Upgrade from 21.12 to 22.12

Jira: CGTS-35774
Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Co-Authored-By: Chris Friesen <chris.friesen@windriver.com>
Change-Id: Ia62ec0d9fee6c8cf4d7d77dfcbfc94124659e4e2
(cherry picked from commit 90fc4e1debd764bff5b3e813d61e45dc503f2378)