Update debian ceph pkg to use git commits
The Debian packaging has been changed to reflect all the
git commits under the directory, and not just the commits
to the metadata folder.

Package builder concatenates the number of revisions added since
the one defined in the BASE_SRCREV to the name of the .deb file
generated. Thus, any new submission to the code in these directories
increment this revision counter in the filename.

Test Plan:
PASS: Created new code submissions to package ceph and verify that
the version in the package name was incremented.
PASS: Created new code submissions to package parted and verify that
the version in the package name was incremented.
PASS: Created new code submissions to package trident-installer and
verify that the version in the package name was incremented.

JIRA: CGTS-45037
Story: 2010550
Task: 47490

(cherry-picked from 4bc42115a6c39062292867e0ebedbf22eae81517)

Signed-off-by: Luis Sampaio <luis.sampaio@windriver.com>
Change-Id: Ia362b3ead9d2428c4fcd2b819aa709c79a4973f7