Update debian package revision counter
The Debian packaging has been changed to reflect all the
git commits under the directory, and not just the commits
to the metadata folder.

Also, it was noticed that the dpdk dl_path url was pointing to
/archive/bullseye/dpdk-bullseye.tar.gz and setting debver to
20.11.6-1~deb11u1. This version should be the one officially supported
on bullseye at that time, and stills the one pointed on the
packages.debian page [1]. The page might be outdated now, since
according to the debian dpdk repository this is no longer the version
tagged for bullseye [2].
In fact, the 20.11.6-1~deb11u1 no longer exists [3] as a valid debian
tag. So, pointing the url to /archive/bullseye/dpdk-bullseye.tar.gz is
already downloading the archive of the Debian release 20.11.7-1~deb11u1.
This change fixes the debver to 20.11.7-1~deb11u1 and sets the url to
explicitly download that version, so this mismatch between debver and
downloaded source is no longer a possibility.

[1] https://packages.debian.org/bullseye/dpdk
[2] https://salsa.debian.org/debian/dpdk/-/commits/bullseye
[3] https://salsa.debian.org/debian/dpdk/-/tags?sort=updated_desc&search=debian%2F

TEST PLAN:
PASS - build-pkgs -c -p openvswitch

Jira: CGTS-43856
Story: 2010550
Task: 47479

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: Idaa90e308fc49efe8d5894f6e458bd786742b96e
(Partial cherry-picked from 99801ab09cb3f2881883c366c2e856ab638b413e