Fix use of ceph_mgr_lifecycle_days variable
The change [1] added a syntax issue where it was not
possible to use the variable ceph_mgr_lifecycle_days.

Analyzing the code, it was possible to observe that
the ceph_mgr_lifecycle_days variable is part of the
Config class, not of the ServiceMonitor.

So to fix the problem, just replace the use of
'self' with CONFIG.

Test Plan:
  PASS: AIO-SX fresh install
  PASS: Notice that there are no errors in the
        log /var/log/mgr-restful-plugin.log

Jira: CGTS-46340
Closes-Bug: 2023553

[1]: https://review.opendev.org/c/starlingx/integ/+/885881

Change-Id: Icb46f1589057607e24123b69e9ab44994580585a
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>
(cherry picked from commit 27379674307d3cd24c5440b2917a21258feefb29)
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>