ts2phc: Fix offset for NMEA delays over 0.5 seconds.
The current code of ts2phc assumes that the NMEA RMC message is received
within 0.5 seconds of the pulse the timestamp in the message is refering
to. However, with most receivers NMEA messages are referenced to the
previous pulse. This causes a 1-second error in the measured offset for
receivers with delays over 0.5 seconds.

Add a 0.5 second correction to map the whole interval between pulses to
the preceding pulse.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>

The patch is backport from:
https://lists.nwtime.org/sympa/arc/linuxptp-devel/2024-01/msg00010.html

Change-Id: Ib52615a69109bd3b1144bba4facf6b96f1c77cf4
Signed-off-by: Jackie Huang <jackie.huang@windriver.com>